generic file structure
WIP need to refine GPt.csv to make sure the data structure and generic data matches whats needed then begin data collection through various sources. Blogs, Articles, AI Generation, as well as other various methods that will be listed on the about us page of the project for proper acredation as well as a button at the button the pull up the list of sources. Apologies for typos its late